move gamble to a different file
